requirements.txt
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
circuits==3.2
cryptography==2.0.3
enum34==1.1.6
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
oauth2client==4.1.2
paypalrestsdk==1.13.1
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
pyOpenSSL==17.2.0
python-twitch-client==0.3.0
requests==2.18.4
rsa==3.4.2
selenium==3.5.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
